+++ src/feature/editPost/ui/EditPost.tsx
@@ -1,0 +1,71 @@
+import {
+  useDialogActions,
+  usePostActions,
+  usePosts,
+  useSelectedPost,
+  useShowEditDialog,
+} from '@/shared/model/store';
+import {
+  Button,
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+  Input,
+  Textarea,
+} from '@/shared/ui';
+
+export const EditPost = () => {
+  const posts = usePosts();
+  const selectedPost = useSelectedPost();
+  const { setPosts, setSelectedPost } = usePostActions();
+
+  const showEditDialog = useShowEditDialog();
+  const { setShowEditDialog } = useDialogActions();
+
+  const updatePost = async () => {
+    try {
+      if (!selectedPost) return;
+
+      const response = await fetch(`/api/posts/${selectedPost.id}`, {
+        method: 'PUT',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(selectedPost),
+      });
+
+      const data = await response.json();
+
+      setPosts(posts.map((post) => (post.id === data.id ? data : post)));
+      setShowEditDialog(false);
+    } catch (error) {
+      console.error('게시물 업데이트 오류:', error);
+    }
+  };
+
+  return (
+    <Dialog
+      open={showEditDialog}
+      onOpenChange={setShowEditDialog}
+    >
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>게시물 수정</DialogTitle>
+        </DialogHeader>
+        <div className='space-y-4'>
+          <Input
+            placeholder='제목'
+            value={selectedPost?.title || ''}
+            onChange={(e) => setSelectedPost({ ...selectedPost, title: e.target.value })}
+          />
+          <Textarea
+            rows={15}
+            placeholder='내용'
+            value={selectedPost?.body || ''}
+            onChange={(e) => setSelectedPost({ ...selectedPost, body: e.target.value })}
+          />
+          <Button onClick={updatePost}>게시물 업데이트</Button>
+        </div>
+      </DialogContent>
+    </Dialog>
+  );
+};
